+++ bookkeeper/view/test.py
@@ -1,0 +1,6 @@
+import sys
+# print(sys.path)
+
+from ..models import *
+
+print('ok')
