+++ bookkeeper/view/view_budget.py
@@ -1,16 +1,7 @@
 import sys
 from PySide6 import QtWidgets
 from PySide6.QtCore import Qt
-
-
-def set_data(table:QtWidgets.QTableWidget, data:list[list[str]]):
-    for i, row in enumerate(data):
-        for j, x in enumerate(row):
-            table.setItem(
-                i, j,
-                QtWidgets.QTableWidgetItem(x.capitalize())
-            )
-
+import utils
 
 
 class BudgetTableWidget(QtWidgets.QWidget):
@@ -34,7 +25,7 @@ class BudgetTableWidget(QtWidgets.QWidget):
         self.header.setSectionResizeMode(
             2, QtWidgets.QHeaderView.ResizeToContents)
 
-        set_data(self.budget_table, [['0', '100'],
+        utils.set_data(self.budget_table, [['0', '100'],
                                 ['0', '200'],
                                 ['0', '1000'] ])
